test: refactor common mock DNS test code
(cherry picked from commit 42db85226c3432247ddb801f6ddd3bab020820bc)
